+++ docs/changelog.rst
@@ -1,12 +1,21 @@
 Changelog
 =========
 
+v0.3 (4/10/20)
+**************
+* Several bug fixes here and there
+* Added markers & jumping to parts of code
+* Added long notes
+* Added forever loops
+* Added waits
+* Compiler launcher now shows the version as well (Use ``python eng.py -v``)
+
 v0.2 (3/10/20)
 **************
-* added documentation
-* added repeat loops
-* added adding and subtracting to a variable
-* added manual error-throwing
+* Added documentation
+* Added repeat loops
+* Added adding and subtracting to a variable
+* Added manual error-throwing
 
 v0.1 (2/10/20)
 **************